+++ io-classes/src/Control/Monad/Class/MonadThrow.hs
@@ -16,7 +16,7 @@ module Control.Monad.Class.MonadThrow
   ( MonadThrow (..)
   , MonadCatch (..)
   , MonadMask (..)
-  , MonadMaskingState (..)
+  , MonadMaskingState
   , MonadEvaluate (..)
   , MaskingState (..)
   , Exception (..)
@@ -193,22 +193,25 @@ data ExitCase a
 --
 class MonadCatch m => MonadMask m where
 
-  {-# MINIMAL mask, uninterruptibleMask #-}
+  {-# MINIMAL mask,
+              uninterruptibleMask,
+              getMaskingState,
+              interruptible #-}
+
   mask, uninterruptibleMask :: ((forall a. m a -> m a) -> m b) -> m b
 
   mask_, uninterruptibleMask_ :: m a -> m a
   mask_                action = mask                $ \_ -> action
   uninterruptibleMask_ action = uninterruptibleMask $ \_ -> action
 
-
-class MonadMask m => MonadMaskingState m where
-  {-# MINIMAL getMaskingState, interruptible #-}
   getMaskingState :: m MaskingState
   interruptible   :: m a -> m a
-  allowInterrupt  :: m ()
 
+  allowInterrupt  :: m ()
   allowInterrupt = interruptible (return ())
 
+class MonadMask m => MonadMaskingState m
+{-# DEPRECATED MonadMaskingState "Use MonadMask instead" #-}
 
 
 -- | Monads which can 'evaluate'.
@@ -254,10 +257,11 @@ instance MonadMask IO where
   uninterruptibleMask  = IO.uninterruptibleMask
   uninterruptibleMask_ = IO.uninterruptibleMask_
 
-instance MonadMaskingState IO where
   getMaskingState = IO.getMaskingState
   interruptible   = IO.interruptible
   allowInterrupt  = IO.allowInterrupt
+
+instance MonadMaskingState IO
 
 instance MonadEvaluate IO where
   evaluate = IO.evaluate
@@ -320,6 +324,11 @@ instance MonadMask m => MonadMask (ReaderT r m) where
     ReaderT $ \e -> uninterruptibleMask $ \u -> runReaderT (a $ q u) e
       where q :: (m a -> m a) -> ReaderT e m a -> ReaderT e m a
             q u (ReaderT b) = ReaderT (u . b)
+
+  getMaskingState = lift getMaskingState
+  interruptible a =
+    ReaderT $ \e -> interruptible (runReaderT a e)
+  allowInterrupt  = lift allowInterrupt
 
 instance (Monad m, MonadEvaluate m) => MonadEvaluate (ReaderT r m) where
   evaluate = lift . evaluate